+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2019-05-24
+
+### Added
+- Added repeating gradient utilities, both linear and radial (see `README` for more info)
 
 ### Changed
 - Values that are optional in `linear-gradient()` and `radial-gradient()` (direction, shape, size, and position) are now omitted when the value is the same as the CSS-defined default value
@@ -66,7 +69,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-gradients/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-gradients/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/benface/tailwindcss-gradients/compare/v2.0.1...v2.1.0
 [2.0.1]: https://github.com/benface/tailwindcss-gradients/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/benface/tailwindcss-gradients/compare/v2.0.0-beta.2...v2.0.0
 [2.0.0-beta.2]: https://github.com/benface/tailwindcss-gradients/compare/v2.0.0-beta.1...v2.0.0-beta.2